Fix sorting problems
Due to using an insertion sort algorithm without a linear data structure (we have tree-dependencies), the sorting could be false somethimes. By using a topological sorting algorithm that operates on a graph we can fix the issues. Secondly the SortDogusByInvertedDependency returned the initial unsorted dogu array in reverse. This one is fixed aswell.