#53 Extract validateDogu function into dogu method, remove mass check
The validate function was not located in the best possible location,
that was: during parsing dogus from strings.

That might interfere with parsing otherwise empty dogus which would then
be possibly considered invalid. Instead, this commit extracts this very
method onto the dogu struct so developers can call this function
independent of the parsing method.

This commit removes the respective bulk dogu counterpart because YAGNI